recieptreader/requirements.txt
imutils==0.5.4
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.13
scipy==1.14.1